+++ examples/wrap-example/index.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import VueLoading from '../../src/vuex-loading';
+import VueLoading from '../../src/vue-loading';
 
 import main from './main.vue';
 
@@ -7,7 +7,10 @@ Vue.use(VueLoading);
 
 new Vue({
   el: '#app',
-  vueLoading: new VueLoading({ registerComponents: false }),
+  vueLoading: new VueLoading({
+    registerComponents: false,
+    componentName: 'my-spinner'
+  }),
   render: function(createElement) {
     return createElement(main);
   }